Workaround for the dreaded memory leak.
There is a bug in pango.create_layout() which allocates memory that is supposed to be released using g_object_unref but under the python API there is no way to release it. It _should_ be released at the same time as the layout structure or at least provide a wrapper for g_object_unref. Instead this patch creates a layout in constructor which is reused and updated instead of recreating a new layout when needed. It has both pros and cons.